feat: Add networking validation
And license headers
