+++ pkg/userdata/errors.go
@@ -1,3 +1,7 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
 package userdata
 
 import "errors"
@@ -14,4 +18,8 @@ var (
 	// Services
 	ErrUnsupportedCNI     = errors.New("unsupported CNI driver")
 	ErrInvalidTrustdToken = errors.New("trustd token is invalid")
+
+	// Networking
+	ErrBadAddressing  = errors.New("invalid network device addressing method")
+	ErrInvalidAddress = errors.New("invalid network address")
 )
